requirements.txt
adjustText==1.1.1
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beartype==0.18.4
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
coverage==7.4.4
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
docutils==0.21.1
executing==2.0.1
fastjsonschema==2.19.1
fiona==1.9.6
fonttools==4.51.0
fqdn==1.5.1
geopandas==0.14.3
griffe==0.44.0
h11==0.14.0
htmltools==0.5.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
keyring==25.1.0
kiwisolver==1.4.5
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mizani==0.11.1
more-itertools==10.2.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nbsphinx==0.9.3
nest-asyncio==1.6.0
nh3==0.2.17
nodeenv==1.8.0
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
numpydoc==1.7.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandas-stubs==2.2.1.240316
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.0
plotnine==0.13.4
pluggy==1.4.0
plum-dispatch==2.3.3
polygon-api-client==1.13.5
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pyparsing==3.1.2
pyproj==3.6.1
pyproject_hooks==1.0.0
pyright==1.1.356
pytest==8.1.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.1
qtconsole==5.5.1
QtPy==2.4.1
quartodoc==0.7.2
questionary==2.0.1
readme_renderer==43.0
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
ruff==0.4.1
scikit-learn==1.4.2
scikit-misc==0.3.1
scipy==1.13.0
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
setuptools==69.5.1
shapely==2.0.4
shiny==0.8.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sphobjinv==2.3.1
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.4.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.3
twine==5.0.0
types-python-dateutil==2.9.0.20240316
types-pytz==2024.1.0.20240417
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.29.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wheel==0.43.0
widgetsnbextension==4.0.10
zipp==3.18.1
